+++ packages/orq-rc/src/models/operations/listdatasets.ts
@@ -156,7 +156,7 @@ export const ListDatasetsData$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-05-25T15:04:44.027Z",
+    "2026-05-25T16:12:11.970Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {